feat: Implement authentication headers for API requests
This commit introduces authentication headers to all API requests made by the `EventsService` and `ClubsService`.

The `useAuth` composable now provides a `getAuthHeader` function that generates the necessary authentication headers based on the stored authentication token. This function is utilized in both services to ensure that all requests are properly authenticated.  A content type parameter was added to the getAuthHeader function to allow for multipart form data

The previous hardcoded authentication details and console logs have been removed from `EventsService` and `ClubsService` for security and code cleanliness.

This change enhances the security of the application by ensuring that only authenticated users can access protected resources.